consolidate plot, plotStratum, plot Trees, plot Error, SamplerState, device
