+++ src/NatCruise.Core/Data/Abstractions/IDeviceDataservice.cs
@@ -1,0 +1,12 @@
+﻿using NatCruise.Models;
+using System.Collections.Generic;
+
+namespace NatCruise.Data
+{
+    public interface IDeviceDataservice : IDataservice
+    {
+        IEnumerable<Device> GetDevices();
+
+        IEnumerable<Device> GetOtherDevices();
+    }
+}
