+++ src/NatCruise.Core/Data/Abstractions/IPlotErrorDataservice.cs
@@ -1,0 +1,10 @@
+﻿using NatCruise.Models;
+using System.Collections.Generic;
+
+namespace NatCruise.Data
+{
+    public interface IPlotErrorDataservice : IDataservice
+    {
+        IEnumerable<PlotError> GetPlotErrors(string unit, int plotNumber);
+    }
+}
