+++ src/NatCruise.Core/Data/Abstractions/ISamplerStateDataservice.cs
@@ -1,17 +1,14 @@
 ﻿using System.Collections.Generic;
-using NatCruise.Cruise.Models;
-using NatCruise.Data;
+using NatCruise.Models;
 
-namespace NatCruise.Cruise.Data
+namespace NatCruise.Data
 {
-    public interface ISampleInfoDataservice : IDataservice
+    public interface ISamplerStateDataservice : IDataservice
     {
         void CopySamplerStates(string deviceFrom, string deviceTo);
         SamplerState GetSamplerState(string stratumCode, string sampleGroupCode, string deviceID);
-        //Device GetCurrentDevice();
-        IEnumerable<Device> GetDevices();
-        IEnumerable<Device> GetOtherDevices();
-        SamplerInfo GetSamplerInfo(string stratumCode, string sampleGroupCode);
+
+
         SamplerState GetSamplerState(string stratumCode, string sampleGroupCode);
         void UpsertSamplerState(SamplerState samplerState);
