+++ src/NatCruise.Core/NatCruise.Core.csproj
@@ -11,6 +11,7 @@
 		<PackageReference Include="Microsoft.AppCenter.Analytics" Version="4.1.0" />
 		<PackageReference Include="Microsoft.AppCenter.Crashes" Version="4.1.0" />
 		<PackageReference Include="Prism.Core" Version="8.0.0.1909" />
+		<PackageReference Include="FMSC.Sampling" Version="2.0.1" />
 	</ItemGroup>
 
 	<ItemGroup Condition="$(UseCruiseDALProjRefs) == 'True'">
