+++ src/NatCruise.Cruise/Services/TreeBasedTallyService.cs
@@ -6,6 +6,7 @@ using NatCruise.Cruise.Models;
 using NatCruise.Data;
 using NatCruise.Models;
 using NatCruise.Navigation;
+using NatCruise.Sampling;
 using System;
 using System.Threading.Tasks;
